+++ CHANGELOG.md
@@ -5,3 +5,5 @@
 - [feat:support webclient and gateway report call metrics](https://github.com/Tencent/spring-cloud-tencent/pull/942)
 - [docs:prevent the release of the final version of the sdk.](https://github.com/Tencent/spring-cloud-tencent/pull/945)
 - [fix: fix nacos CircuitBreaker disable bug.](https://github.com/Tencent/spring-cloud-tencent/pull/947)
+- [feature: add config for customized local port.](https://github.com/Tencent/spring-cloud-tencent/pull/956)
+- [refactor:refactor reactor code in router module.](https://github.com/Tencent/spring-cloud-tencent/pull/958)